Fixed the DEBUG build, quoted git hashstring error messages.
There were two problem that hindered Debug builds:
- Implicit linking nonsense in tbbmalloc (removed that).
- Wrong path for .pdb files.
Warning: on Visual Studio, cmake must be called with the DEBUG option set to ON
to be able to perform a Debug build. The CMake invocation determines the kind of
build (Release or Debug) and you cannot run the other kind of builds. If you
want both types of build, you *must* use two build trees.